requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

h5py==3.8.0
numpy==2.1.3
pandas==2.2.3
pytorch-lightning==2.5.1
PyYAML==6.0.2
transformers==4.29.2
x-transformers==2.2.11
